enable score based selection
